Add businesses to control panel
Note: This implementation consists of a lot of boilerplate code
copy-pasted from the ListingAdmin component. These two components are
therefore ripe for abstraction. We should for example consider moving
image uploads (the uploadLogo function) into a service of some kind. We
could also benefit from having a folder with utils where we can move
the validateLink function. The vuex model for businesses is also very
similar to that of listings.